Mitigate k8s CRD race condition
If Kubernetes Custom Resource Definitions and their corresponding
Custom Resources are applied from a single manifest, there is a race
condition where the system may try to instantiate the CR before the
CRD is fully processed. This issue is known and will not be fixed.

Mitigation strategies are as follows:

  1. Split CRDs and CRs into separate manifests and apply them
     sequentially.
  2. Apply a joint manifest. In case of failure, wait for the CRDs
     to become established and apply the manifest again.

In both cases `kubectl wait` can be employed to ensure that the CRDs
are fully established.

For simplicity, this commit uses strategy 2.