Finish transition to GitOps
Cluster bootstrapping is now clean in that all resources that did not
come from a Git deploy/ directory are deleted when no longer needed.
As a result, the cluster only contains Flux-controlled resources.

In particular, the TLS secret used for GitOps is itself versioned and
can be rotated in pure GitOps fashion.